Fix linux rendering issues (#2)
* Debugging some lighting issues on Linux

* Lots of refactoring. Lights still not working in linux

* Fixed a wrong named uniform. More refactoring

* More refactoring

* More refactoring. Re-enabling some disabled stuff

* Use GLuint in certain OpenGL loops instead of size_t. Use vec4 instead of vec3 for boxes' colors

* Remove extra output

* more more..

* Always a great idea to push a bunch of commits in a debug branch

* Aesthetic refactoring

* Use local GLM only

* Last changes before merge

+++ src/framework/3d/texture.cpp
@@ -34,7 +34,7 @@ namespace Framework {
         unsigned char* image;
 
         glBindTexture(GL_TEXTURE_CUBE_MAP, textureID);
-        for(size_t i = 0; i < m_faces.size(); i++)
+        for(GLuint i = 0; i < m_faces.size(); i++)
         {
             image = stbi_load(m_faces[i], &width, &height, &numComponents, 3);
             glTexImage2D(GL_TEXTURE_CUBE_MAP_POSITIVE_X + i, 0, GL_SRGB, width, height, 0, GL_RGB, GL_UNSIGNED_BYTE, image);
